Update About page
I also noticed that there is a slight 'jitter; when switching to
and from the Blog page due to a difference in screen/margin width
of 15 px. I don't know where this difference in width is coming
from. It looks it has something to do with the <ul> since
removing it eliminates the jitter.